+++ front/src/app/modules/log-dialog/log-dialog.module.ts
@@ -5,7 +5,6 @@ import { MatDialogModule } from "@angular/material/dialog";
 import { FlexModule } from "@angular/flex-layout";
 import { MatButtonModule } from "@angular/material/button";
 import { OpenLogDialogService } from "./open-log-dialog.service";
-import { MatExpansionModule } from '@angular/material/expansion';
 import { TranslateModule } from '@ngx-translate/core';
 
 
@@ -19,7 +18,6 @@ import { TranslateModule } from '@ngx-translate/core';
     MatDialogModule,
     FlexModule,
     MatButtonModule,
-    MatExpansionModule,
     TranslateModule
   ],
   providers: [