+++ front/src/app/modules/log-dialog/log-dialog/log-dialog.component.ts
@@ -1,5 +1,7 @@
 import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material/dialog";
+import { IServiceLogs } from "../../shell/interfaces/container";
+import { stripRegistryHost } from "../image-reference";
 
 @Component({
   templateUrl: './log-dialog.component.html',
@@ -7,10 +9,14 @@ import { MAT_DIALOG_DATA, MatDialogRef } from "@angular/material/dialog";
 })
 export class LogDialogComponent {
 
+  readonly imageReference: string;
+
   constructor(
-    @Inject(MAT_DIALOG_DATA) public data: Record<string, string>,
+    @Inject(MAT_DIALOG_DATA) public data: IServiceLogs,
     private readonly dialogRef: MatDialogRef<LogDialogComponent>,
-  ) {}
+  ) {
+    this.imageReference = stripRegistryHost(this.data?.image);
+  }
 
   close() {
     this.dialogRef.close();